Utility, vaapi: Move definition of CheckAndDestroyVAbuffer to common utility header
